replace progression by progress
