+++ app/src/components/ProgressBar.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react'
-import useOngoingTransferProgression from '@/hooks/useOngoingTransferProgression'
+import useOngoingTransferProgress from '@/hooks/useOngoingTransferProgress'
 import { StoredTransfer } from '@/models/transfer'
 import { SnowbridgeStatus } from '@/models/snowbridge'
 import { Direction } from '@/services/transfer'
@@ -18,7 +18,7 @@ const ProgressBar: FC<ProgressBarProps> = ({
   bridgeStatus,
   outlinedProgressBar,
 }) => {
-  const progress = useOngoingTransferProgression(transfer, direction, bridgeStatus)
+  const progress = useOngoingTransferProgress(transfer, direction, bridgeStatus)
 
   return (
     <>
@@ -26,20 +26,20 @@ const ProgressBar: FC<ProgressBarProps> = ({
         <div
           className={cn(
             'mb-4 h-2 overflow-hidden rounded-full bg-turtle-secondary-light',
-            progression <= 0 && 'animate-pulse',
+            progress <= 0 && 'animate-pulse',
           )}
         >
           <div
             className="h-full rounded-full border border-turtle-secondary-dark bg-turtle-secondary transition-all duration-700 ease-in-out"
-            style={{ transform: `translateX(-${100 - (progression || 0)}%)` }}
+            style={{ transform: `translateX(-${100 - (progress || 0)}%)` }}
           />
         </div>
       ) : (
         <div className="mb-4 h-2 rounded-full border border-turtle-secondary bg-white">
-          {progression > 0 && (
+          {progress > 0 && (
             <div
               className="-ml-[1px] -mt-[1px] h-2 rounded-full border border-turtle-secondary-dark bg-turtle-secondary"
-              style={{ width: `${progression}%` }}
+              style={{ width: `${progress}%` }}
             />
           )}
         </div>
